// apps/backend/api/src/types/requests.ts
export interface components {
  schemas: {
// ...
    Denial: {
      /**
       * Format: date-time
       * @description Timestamp the record was created in ISO 8601 format with time zone (e.g., "2024-12-09T15:30:00Z").
       */
      createdAt?: string;
      /**
       * Format: date-time
       * @description Date denied in ISO 8601 format with time zone (e.g., "2024-12-09T15:30:00Z").
       */
      date_denied: string;
      id?: number;
      is_final?: boolean;
      read?: boolean;
      /** @description Denial Reason. */
      reason: string;
      request_id: number;
      user_id?: string;
      /**
       * Format: date-time
       * @description Timestamp the record was updates in ISO 8601 format with time zone (e.g., "2024-12-09T15:30:00Z").
       */
      updatedAt?: string;
    };
// ...
  };
// ...
}